Update faq.md
There are a few changes here. 

A grammar checker might have told you to write "Rails need" because it mistakenly thought of "Rails" as a plural noun, but in the case of Ruby on Rails, it's a singular noun.  For example these 2 sentences are both correct, but use the word "Rails" differently:

* Rails _are_ used in making railroads and roller coasters.
* Rails _is_ an awesome web-development framework written in Ruby.

In the rest of the sentence I tried to rewrite it in the most natural English possible, which meant putting a preposition at the end.  (Many English teachers make the claim that doing so is bad grammar, and I would be happy to explain why I believe they are badly mistaken, but it's a big topic so I won't do that unless you want to hear more about that topic.)